+++ avalon-modules/avalon-auth/avalon-auth-repository/src/main/java/io/github/lishangbu/avalon/auth/repository/RoleRepository.java
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface RoleRepository
-    extends ListCrudRepository<Role, Integer>, ListPagingAndSortingRepository<Role, Integer> {
+    extends ListCrudRepository<Role, Long>, ListPagingAndSortingRepository<Role, Long> {
 
   Optional<Role> findByCode(String code);
 }
