+++ avalon-modules/avalon-dataset/avalon-dataset-repository/src/main/java/io/github/lishangbu/avalon/dataset/repository/BerryRepository.java
@@ -11,4 +11,4 @@ import org.springframework.stereotype.Repository;
  * @since 2025/5/22
  */
 @Repository
-public interface BerryRepository extends ListPagingAndSortingRepository<Berry, Integer> {}
+public interface BerryRepository extends ListPagingAndSortingRepository<Berry, Long> {}
