+++ avalon-modules/avalon-dataset/avalon-dataset-repository/src/main/java/io/github/lishangbu/avalon/dataset/repository/TypeDamageRelationRepository.java
@@ -12,4 +12,4 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface TypeDamageRelationRepository
-    extends ListPagingAndSortingRepository<TypeDamageRelation, Integer> {}
+    extends ListPagingAndSortingRepository<TypeDamageRelation, Long> {}
